# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  handlers.BlogRequest:
    properties:
      authorid:
        type: integer
      score:
        type: number
      title:
        type: string
    type: object
...
  handlers.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  models.Blog:
    properties:
      authorID:
        type: integer
      createdDate:
        type: string
      id:
        type: integer
      score:
        type: number
      title:
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Practice Go API using the Standard Library and Postgres
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog Service API
  version: "1.0"
paths:
...
  /blog/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Blog by ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete Blog
      tags:
      - blog
    get:
      consumes:
      - application/json
      description: Read Blog by ID
      parameters:
      - description: Blog Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Read Blog
      tags:
      - blog
    put:
      consumes:
      - application/json
      description: Update Blog by ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Blog to Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update Blog
      tags:
      - blog
...
  /user:
    get:
      consumes:
      - application/json
      description: List All Users
      parameters:
      - description: query param
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Creates a User
      parameters:
      - description: User to Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create User
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete User
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Read User by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Read User
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update User by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User to Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update User
      tags:
      - user
swagger: "2.0"